Enable torch.load in dataset_pyg.py
It seems like Torch has recently changed its load function to default to "weights_only=True". This causes errors in the current version of this library.  I am not an expert, and I don't know if my solution is safe, but this was my workaround to get your code running.